Time: 3 ms (86.04%) | Memory: 71.8 MB (67.25%) - LeetSync

+++ 1445-number-of-sub-arrays-of-size-k-and-average-greater-than-or-equal-to-threshold/number-of-sub-arrays-of-size-k-and-average-greater-than-or-equal-to-threshold.java
@@ -12,7 +12,7 @@ class Solution {
                 low++;
             }
             if(high-low+1==k){
-                double ans=(double)avg/k;
+                int ans=avg/k;
 
                 if(ans>=threshold){
                     count++;
